Fixed "early access" detection.
A brief history: tags used to be a valid way to detect early access. If
the tag was present, it was EA, otherwise it was not. Later, the early
access tag appeared twice on games, one browseable and one not. If the
non-browseable tag was present, the game was actually in EA, otherwise
it was a game that used to be in EA but was no longer.

Later, the duplicate EA tags were reduced to a single tag, but whether
it was the browseable or not was pseudo-random, and in most cases it was
the browseable one, meaning most games were no longer detected as EA.
Tags were no longer a reliable mechanism for detecting EA. However,
genres now are.